# swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"FieldErrors": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"value": {},
						"message": {
							"type": "string"
						}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"ErrorDetails": {
				"anyOf": [
					{
						"properties": {
							"folderName": {
								"type": "string"
							},
							"userId": {
								"type": "string"
							},
							"folderId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"imageUrl": {
								"type": "string"
							},
							"imageId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"userId": {
								"type": "string"
							},
							"challengeId": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"longitude": {
								"type": "number",
								"format": "double"
							},
							"latitude": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"reason": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"searchKeyword": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"$ref": "#/components/schemas/FieldErrors"
					},
					{
						"properties": {
							"extension": {
								"type": "string"
							}
						},
						"type": "object"
					},
					{
						"properties": {
							"imageId": {
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							"folderId": {
								"type": "string"
							}
						},
						"type": "object"
					}
				],
				"nullable": true
			},
			"ITsoaErrorResponse": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"properties": {
							"data": {
								"$ref": "#/components/schemas/ErrorDetails"
							},
							"reason": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"reason",
							"errorCode"
						],
						"type": "object"
					},
					"success": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ITsoaSuccessResponse__folder_id-string--image_text-string__": {
				"properties": {
					"resultType": {
						"type": "string"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"success": {
						"properties": {
							"image_text": {
								"type": "string"
							},
							"folder_id": {
								"type": "string"
							}
						},
						"required": [
							"image_text",
							"folder_id"
						],
						"type": "object"
					}
				},
				"required": [
					"resultType",
					"error",
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"sessionAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "connect.sid"
			}
		}
	},
	"info": {
		"title": "sweepic-server",
		"version": "1.0.0",
		"description": "Sweepic 프로젝트",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Sweepic local server"
		},
		{
			"url": "http://3.37.137.212:3000",
			"description": "Sweepic server"
		}
	],
	"paths": {
...
		"/memo/text-format/folders/{folderId}": {
			"patch": {
				"operationId": "UpdateFolderOCR",
				"responses": {
					"200": {
						"description": "폴더 업데이트 및 텍스트 변환",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaSuccessResponse__folder_id-string--image_text-string__"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "SUCCESS",
											"error": null,
											"success": {
												"folder_id": "1",
												"image_text": "이번 수업 시간은 사회 과학 시간이다."
											}
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "잘못된 요청 데이터",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "SRH-400",
												"reason": "입력 데이터가 유효하지 않습니다.",
												"data": {
													"reason": "folder_ID가 필요합니다."
												}
											}
										}
									},
									"Example 2": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "SRH-400",
												"reason": "입력 데이터가 유효하지 않습니다.",
												"data": {
													"reason": "base64_image가 필요합니다."
												}
											}
										}
									},
									"Example 3": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "SRH-400",
												"reason": "입력 데이터가 유효하지 않습니다.",
												"data": {
													"reason": "user_id가 필요합니다."
												}
											}
										}
									},
									"Example 4": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "PHO-400",
												"reason": "사진 데이터가 유효하지 않습니다.",
												"data": {
													"reason": "올바른 Base64 이미지 형식이 아닙니다."
												}
											}
										}
									},
									"Example 5": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "PHO-400",
												"reason": "사진 데이터가 유효하지 않습니다.",
												"data": {
													"reason": "텍스트를 추출할 사진이 없습니다."
												}
											}
										}
									},
									"Example 6": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "PHO-400",
												"reason": "사진 데이터가 유효하지 않습니다.",
												"data": {
													"reason": "이미지에서 텍스트를 찾지 못하였습니다."
												}
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않은 데이터 조회 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ITsoaErrorResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"resultType": "FAIL",
											"success": null,
											"error": {
												"errorCode": "FOL-404",
												"reason": "해당 폴더를 찾을 수 없습니다.",
												"data": {
													"folderId": "1"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "기존 폴더에 이미지를 추가하고, OCR 텍스트를 업데이트하는 API입니다.",
				"summary": "폴더 업데이트 및 OCR 수행",
				"tags": [
					"memo-ai"
				],
				"security": [],
				"parameters": [
					{
						"description": "업데이트할 폴더의 ID",
						"in": "path",
						"name": "folderId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"base64_image": {
										"type": "string",
										"format": "binary",
										"description": "OCR 처리를 위한 이미지 업로드"
									}
								}
							}
						}
					}
				}
			}
		},
...
	}
}